Refactor and init_ifc section
In case of simulator initialization function is added. And in INTERFACE some refactoring is applied. The peripheral specific inf functions has been separated to different source files.
