+++ .gitignore
@@ -1,11 +1,6 @@
 .vscode/settings.json
-APPLICATION/HARDWARE/OBJ/hw_led_handler.o
-APPLICATION/INTERFACE/OBJ/main_ifc.o
-APPLICATION/LED_HANDLER/OBJ/led_handler.o
-APPLICATION/SIMULATOR/OBJ/sim_led_handler.o
-APPLICATION/TARGET/application.exe
-HARDWARE/OBJ/hw_led_handler.o
-INTERFACE/OBJ/main_ifc.o
-LED_HANDLER/OBJ/led_handler.o
-SIMULATOR/OBJ/sim_led_handler.o
-TARGET/application.exe
+TARGET/
+APPLICATION/OBJ/
+HARDWARE/OBJ/
+INTERFACE/OBJ/
+SIMULATOR/OBJ/
