+++ APPLICATION/SRC/led_handler.c
@@ -3,8 +3,11 @@
 
 int main( void )
 {
+
+  //Initialize section:
+  init_ifc();
+
   //Inifinit loop:
-  
   led_on_ifc();
   led_off_ifc();
