+++ APPLICATION/sources.mk
@@ -1,0 +1,3 @@
+GENERAL_SRC_LIST+=./APPLICATION/SRC/led_handler.c
+
+INCLUDE+=-I./APPLICATION/HEADER/
