+++ INTERFACE/HEADER/common_defines.h
@@ -1,0 +1,14 @@
+#ifndef _COMMON_DEFINES_
+#define _COMMON_DEFINES_
+
+typedef enum{
+
+    INIT_FN = 0,
+    FN_LED_ON,
+    FN_LED_OFF,
+    FN_MAX_VALUE
+
+}FN_PTR_LIST;
+
+
+#endif /* _COMMON_DEFINES_  */
