+++ INTERFACE/HEADER/init_ifc.h
@@ -1,0 +1,9 @@
+#ifndef _INIT_IFC_
+#define _INIT_IFC_
+
+#include "common_defines.h"
+#include "main_ifc.h"
+
+void init_ifc( void );
+
+#endif /* _INIT_IFC_ */
