+++ INTERFACE/HEADER/led_handler_ifc.h
@@ -1,0 +1,10 @@
+#ifndef _LED_HANDLER_IFC_
+#define _LED_HANDLER_IFC_
+
+#include "common_defines.h"
+#include "main_ifc.h"
+
+void led_on_ifc( void );
+void led_off_ifc( void );
+
+#endif /* _LED_HANDLER_IFC_ */
