+++ INTERFACE/HEADER/main_ifc.h
@@ -1,6 +1,7 @@
 #ifndef _MAIN_IFC_H_
 #define _MAIN_IFC_H_
 
+#include "common_defines.h"
 #include "connector.h"
 
 /*
@@ -11,16 +12,8 @@ According to the used BUILD variant this layer makes connection between a phisyc
 or simulator.
 */
 
-typedef enum{
 
-    FN_LED_ON = 0,
-    FN_LED_OFF,
-    FN_MAX_VALUE
-
-}FN_PTR_LIST;
-
-void led_on_ifc( void );
-void led_off_ifc( void );
-void ERROR_HANDLER( const char* fn_name , const char* error_msg );
+void execute_function( FN_PTR_LIST );
+void ERROR_HANDLER( const char* , const char* );
 
 #endif /* _MAIN_IFC_H_ */
