+++ INTERFACE/SRC/init_ifc.c
@@ -1,0 +1,8 @@
+#include "init_ifc.h"
+
+void init_ifc( void )
+{
+    execute_function( INIT_FN );
+}
+
+
