+++ INTERFACE/sources.mk
@@ -1,3 +1,5 @@
-GENERAL_SRC_LIST+=./INTERFACE/SRC/main_ifc.c
+GENERAL_SRC_LIST+=./INTERFACE/SRC/main_ifc.c \
+./INTERFACE/SRC/init_ifc.c \
+./INTERFACE/SRC/led_handler_ifc.c
 
 INCLUDE+=-I./INTERFACE/HEADER/
