--- LED_HANDLER/sources.mk
@@ -1,3 +1,0 @@
-GENERAL_SRC_LIST+=./LED_HANDLER/SRC/led_handler.c
-
-INCLUDE+=-I./LED_HANDLER/HEADER/
