+++ SIMULATOR/HEADER/sim_led_handler.h
@@ -1,6 +1,7 @@
 #ifndef _SIM_LED_HANDLER_H_
 #define _SIM_LED_HANDLER_H_
 
+void sim_led_init   ( void );
 void sim_led_on     ( void );
 void sim_led_off    ( void );
