+++ SIMULATOR/SRC/sim_led_handler.c
@@ -1,9 +1,15 @@
 #include <stdio.h>
 #include "sim_led_handler.h"
 
-void (*fn_ptr_list[])( void ) = {   sim_led_on, 
+void (*fn_ptr_list[])( void ) = {   sim_led_init,
+                                    sim_led_on, 
                                     sim_led_off
                                 };
+
+void sim_led_init( void )
+{
+    printf("I'm %s\n", __func__  );
+}
 
 void sim_led_on( void )
 {
